Use Push Btns on 7 & 8 to send Data to ThingSpeak
I put this together really quick and dirty from our other ESP Sketch. It sends values to ThingSpeak.com when 7 or 8 are shorted to Ground! I used it to test out and learn about sending data via ESP/Arduino but if you want to use it, it needs a LOT of refinement and debugging.